feat(exec): add glob support for filenames
Filenames can now be passed with glob patterns.
On Windows the glob patterns need to use the '/' slash and not the
'\'. So e.g. ("**/*.dlt")
Filenames must be passed on the console with "" to avoid the
console itself to search/glob the file patterns.
Main use-case is for windows users.